+++ engine/runtime/src/main/scala/org/enso/compiler/pass/resolve/VectorLiterals.scala
@@ -101,7 +101,7 @@ case object VectorLiterals extends IRPass {
             BindingsMap
               .ResolvedConstructor(
                 ModuleReference.Concrete(module),
-                BindingsMap.Cons("Vector", 1)
+                BindingsMap.Cons("Vector", 1, allFieldsDefaulted = false)
               )
           )
         )